Divide classes into packages ----------------------------------------- 1. duke.command 2. duke.task
