+++ src/main/java/duke/Ui.java
@@ -1,3 +1,7 @@
+package duke;
+
+import duke.task.Task;
+
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.Scanner;
