+++ src/main/java/duke/command/ByeCommand.java
@@ -1,19 +1,20 @@
-public class DoneCommand extends Command {
-    private int index;
+package duke.command;
 
-    public DoneCommand(int index) {
-        this.index = index;
-    }
+import duke.Storage;
+import duke.Ui;
+import duke.task.TaskList;
+
+public class ByeCommand extends Command {
 
     @Override
     public void execute(TaskList tasks, Storage storage, Ui ui) {
-        tasks.finishTask(index);
         String saveFileString = tasks.save();
         storage.save(saveFileString);
+        Ui.endDuke();
     }
 
     @Override
     public boolean isExit() {
-        return false;
+        return true;
     }
 }
