+++ src/main/java/duke/command/DeleteCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.Storage;
+import duke.Ui;
+import duke.task.TaskList;
+
 public class DeleteCommand extends Command {
     private int index;
