+++ src/main/java/duke/command/EmptyCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.Storage;
+import duke.Ui;
+import duke.task.TaskList;
+
 public class EmptyCommand extends Command {
     @Override
     public void execute(TaskList tasks, Storage storage, Ui ui) {
