+++ src/main/java/duke/command/EventCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.Storage;
+import duke.Ui;
+import duke.task.Event;
+import duke.task.TaskList;
 import java.time.LocalDateTime;
 
 public class EventCommand extends Command{
