+++ src/main/java/duke/task/ToDo.java
@@ -1,3 +1,5 @@
+package duke.task;
+
 /**
  * Represents a ToDo, which is the simplest of Tasks.
  */
